# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tokens": {
      "post": {
        "operationId": "TokenController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "summary": "Create a new Token with optional addresses",
        "tags": [
          "tokens"
        ]
      },
      "get": {
        "operationId": "TokenController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all Tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Token"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all Tokens with their addresses",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/{id}": {
      "get": {
        "operationId": "TokenController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Token UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "404": {
            "description": "Token not found"
          }
        },
        "summary": "Get a Token by ID with its addresses",
        "tags": [
          "tokens"
        ]
      },
      "put": {
        "operationId": "TokenController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Token UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Token not found"
          }
        },
        "summary": "Update a Token",
        "tags": [
          "tokens"
        ]
      },
      "delete": {
        "operationId": "TokenController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Token UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Token deleted successfully"
          },
          "404": {
            "description": "Token not found"
          }
        },
        "summary": "Delete a Token",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/{id}/addresses": {
      "post": {
        "operationId": "TokenController_addAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Token UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "TokenAddress added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenAddress"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Token not found"
          },
          "409": {
            "description": "Address already exists on that chain"
          }
        },
        "summary": "Add a new address to a Token",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/{id}/addresses/{addressId}": {
      "put": {
        "operationId": "TokenController_updateAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Token UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "description": "TokenAddress UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokenAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TokenAddress updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenAddress"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Token or TokenAddress not found"
          },
          "409": {
            "description": "Address/chain combination already exists"
          }
        },
        "summary": "Update a TokenAddress",
        "tags": [
          "tokens"
        ]
      },
      "delete": {
        "operationId": "TokenController_removeAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Token UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "description": "TokenAddress UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "TokenAddress deleted successfully"
          },
          "404": {
            "description": "Token or TokenAddress not found"
          }
        },
        "summary": "Delete a TokenAddress from a Token",
        "tags": [
          "tokens"
        ]
      }
    }
  },
  "info": {
    "title": "Arbi Engine API",
    "description": "API documentation for Arbi Engine operations",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTokenAddressDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Contract address on the blockchain"
          },
          "chainId": {
            "type": "number",
            "description": "Blockchain chain ID"
          },
          "decimals": {
            "type": "number",
            "description": "Number of decimals for the token"
          }
        },
        "required": [
          "address",
          "chainId"
        ]
      },
      "CreateTokenDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Token symbol (e.g., USDC, WETH)"
          },
          "name": {
            "type": "string",
            "description": "Full token name"
          },
          "description": {
            "type": "string",
            "description": "Token description"
          },
          "addresses": {
            "description": "Array of token addresses across different chains",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTokenAddressDto"
            }
          }
        },
        "required": [
          "symbol",
          "name"
        ]
      },
      "Token": {
        "type": "object",
        "properties": {}
      },
      "UpdateTokenDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Token symbol (e.g., USDC, WETH)"
          },
          "name": {
            "type": "string",
            "description": "Full token name"
          },
          "description": {
            "type": "string",
            "description": "Token description"
          },
          "addresses": {
            "description": "Array of token addresses across different chains",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTokenAddressDto"
            }
          }
        }
      },
      "TokenAddress": {
        "type": "object",
        "properties": {}
      },
      "UpdateTokenAddressDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Contract address on the blockchain"
          },
          "chainId": {
            "type": "number",
            "description": "Blockchain chain ID"
          },
          "decimals": {
            "type": "number",
            "description": "Number of decimals for the token"
          }
        }
      }
    }
  }
}